Move client-specific methods from base class `RoomCoordinator`
